fix(components): break import cycle introduced by the OPClient collapse
Regression from 28d350f. Retargeting components/store/index.ts at
components/op.ts closed a cycle:

  op.ts -> helpers.ts -> @dynamic/1password/OnePasswordItem.ts
        -> components/store/index.ts -> op.ts

store/index.ts constructs an OPClient at module scope, so entering that cycle
via op.ts runs the constructor before the class binding initialises:

  Error: Cannot access 'OPClient' before initialization

The duplicate store/op.ts had removeUndefinedProperties inlined rather than
imported, which is what kept the cycle open before — collapsing the duplicate
removed that accidental firebreak.

tsc did not catch it: the types are fine, the failure is initialisation order.
It only surfaced when scripts/op-to-bao imported @components/op.ts as the
entry point. Verified against the pushed branch — `import('@components/op.ts')`
throws there and loads cleanly here, as do store/index.ts,
OnePasswordItem.ts and helpers.ts in any order.

Fix moves removeUndefinedProperties into its own dependency-free module that
both helpers.ts and op.ts import, so op.ts no longer reaches helpers.ts at all.
helpers.ts re-exports it, so every existing consumer is untouched.
